Playing with strict order execution in parallel
Seems to be a stupid task, since in current task everything is done
sequentially in two threads. But we still can look at it as at some
estimation od the thread switching using different tqchnique ?
Ot maybe later the task will be changed so there will be some work done
in parallel.